list view cards on activity log

+++ src/main/java/org/tin/oop2_capstone/controllers/LogCardController.java
@@ -1,17 +1,18 @@
 package org.tin.oop2_capstone.controllers;
 
 import javafx.fxml.FXML;
+import javafx.scene.control.Label;
 
-import java.awt.*;
 
 public class LogCardController {
     @FXML Label logNameLabel;
     @FXML Label timeLabel;
     @FXML Label numberLabel;
     @FXML Label unitLabel;
-    @FXML Label kcalLabel;
+    @FXML
+    Label kcalLabel;
 
-    public void setData(String logName, String time, double number, String unit, int kcal){
+    public void setData(String logName, String time, double number, String unit, double kcal){
         logNameLabel.setText(logName);
         timeLabel.setText(time);
         numberLabel.setText(String.valueOf(number));
